Update cmd/viewlogs: refactor table consturucting
