+++ cmd/viewlogs/app/run.go
@@ -5,12 +5,8 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"slices"
 	"strings"
-	"time"
 
-	"github.com/fatih/color"
-	"github.com/rodaine/table"
 	"github.com/spf13/pflag"
 	"go.etcd.io/bbolt"
 
@@ -69,52 +65,14 @@ func run(filename, operation, resource string) error {
 		return errors.New("resource id must be specified")
 	}
 
-	type row struct {
-		ts        int64 // timestamp
-		operation string
-		data      string
-		context   string
-	}
-	var rows []row
-	insert := func(r row) {
-		idx, _ := slices.BinarySearchFunc(rows, r, func(r1, r2 row) int {
-			switch {
-			case r1.ts == r2.ts:
-				return 0
-			case r1.ts < r2.ts:
-				return -1
-			default:
-				return 1
-			}
-		})
-
-		if idx == len(rows) {
-			rows = append(rows, r)
-		} else {
-			rows = append(rows, row{})
-			copy(rows[idx+1:], rows[idx:])
-			rows[idx] = r
-		}
-	}
+	table := newTablePrinter()
 
 	if server {
 		store, err := logs.NewInfoStore(db)
 		if err != nil {
 			return err
 		}
-
-		for _, l := range store.ServerRecord().Logs {
-			var data string
-			if l.Event == logsv1.ServerEvent_SERVER_EVENT_LAUNCHED {
-				data = fmt.Sprintf("addr=%s", l.Addr)
-			}
-			insert(row{
-				ts:        l.Timestamp,
-				operation: formatServerOperation(l.Event),
-				context:   logs.CallerContext(l.Context).String(),
-				data:      data,
-			})
-		}
+		table.insertServerLogs(store.ServerRecord().Logs)
 	}
 
 	if init {
@@ -122,19 +80,11 @@ func run(filename, operation, resource string) error {
 		if err != nil {
 			return err
 		}
-
 		r, err := store.Get(resource)
 		if err != nil {
 			return err
 		}
-		for _, l := range r.Logs {
-			insert(row{
-				ts:        l.Timestamp,
-				operation: formatInitOperation(l.Event),
-				context:   logs.CallerContext(l.Context).String(),
-				data:      "",
-			})
-		}
+		table.insertInitLogs(resource, r.Logs)
 	}
 
 	if acquire {
@@ -142,94 +92,15 @@ func run(filename, operation, resource string) error {
 		if err != nil {
 			return err
 		}
-
 		r, err := store.Get(resource)
 		if err != nil {
 			return err
 		}
-		for i, l := range r.Logs {
-			var data string
-			if i == 0 {
-				data = fmt.Sprintf("Max=%d,", r.Max)
-			}
-			if l.Event == logsv1.AcquisitionEvent_ACQUISITION_EVENT_ACQUIRED {
-				data = fmt.Sprintf("%sN=%d", data, l.N)
-			}
-			insert(row{
-				ts:        l.Timestamp,
-				operation: formatAcquisitionOperation(l.Event),
-				context:   logs.CallerContext(l.Context).String(),
-				data:      data,
-			})
-		}
+		table.insertAcquisitionLogs(resource, r)
 	}
 
 	fmt.Printf("Resource identifier: %q\n\n", resource)
-	tbl := table.New("Time", "Elapsed", "Operation", "Data", "Context(Map->Resource)").
-		WithHeaderFormatter(
-			color.New(color.FgGreen, color.Underline).SprintfFunc(),
-		).
-		WithFirstColumnFormatter(
-			color.New(color.FgYellow).SprintfFunc(),
-		)
-
-	var last time.Time
-	for _, r := range rows {
-		timestamp, elapsed := formatTime(r.ts, &last)
-		tbl.AddRow(timestamp, elapsed, r.operation, r.data, r.context)
-	}
-	tbl.Print()
+	table.print()
 
 	return nil
-}
-
-func formatServerOperation(e logsv1.ServerEvent) string {
-	switch e {
-	case logsv1.ServerEvent_SERVER_EVENT_LAUNCHED:
-		return "server:launched"
-	case logsv1.ServerEvent_SERVER_EVENT_STOPPED:
-		return "server:stopped"
-	default:
-		return e.String()
-	}
-}
-
-func formatInitOperation(e logsv1.InitEvent) string {
-	switch e {
-	case logsv1.InitEvent_INIT_EVENT_STARTED:
-		return "init:started"
-	case logsv1.InitEvent_INIT_EVENT_COMPLETED:
-		return "init:completed"
-	case logsv1.InitEvent_INIT_EVENT_FAILED:
-		return "init:failed"
-	default:
-		return e.String()
-	}
-}
-
-func formatAcquisitionOperation(e logsv1.AcquisitionEvent) string {
-	switch e {
-	case logsv1.AcquisitionEvent_ACQUISITION_EVENT_ACQUIRED:
-		return "acquired"
-	case logsv1.AcquisitionEvent_ACQUISITION_EVENT_RELEASED:
-		return "released"
-	default:
-		return e.String()
-	}
-}
-
-func formatTime(ts int64, last *time.Time) (string, string) {
-	t := time.Unix(0, ts)
-	defer func() {
-		*last = t
-	}()
-	s := t.Format("2006-01-02 15:04:05.999999999")
-	if last.IsZero() {
-		return s, ""
-	}
-	diff := t.Sub(*last)
-	if diff >= 0 {
-		return s, "+" + diff.String()
-	}
-	return s, diff.String()
 }
